refactor: rename WizardModal to LegacyWizardModal
