+++ frontend/src/lib/modals/LegacyWizardModal.svelte
