+++ frontend/src/lib/modals/accounts/RenameSubAccountModal.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import RenameSubAccountAction from "../../components/accounts/RenameSubAccountAction.svelte";
-  import WizardModal from "../WizardModal.svelte";
+  import LegacyWizardModal from "../LegacyWizardModal.svelte";
   import { i18n } from "../../stores/i18n";
   import type { Steps } from "../../stores/steps.state";
   import type { Step } from "../../stores/steps.state";
@@ -16,7 +16,7 @@
   let currentStep: Step | undefined;
 </script>
 
-<WizardModal {steps} bind:currentStep on:nnsClose>
+<LegacyWizardModal {steps} bind:currentStep on:nnsClose>
   <svelte:fragment slot="title"
     >{$i18n.accounts.rename_linked_account}</svelte:fragment
   >
@@ -24,4 +24,4 @@
   <svelte:fragment>
     <RenameSubAccountAction on:nnsClose />
   </svelte:fragment>
-</WizardModal>
+</LegacyWizardModal>
